make update block work and save/update data

+++ src/AirSnitch.Domain/Models/Pm10Particle.cs
@@ -2,11 +2,11 @@ namespace AirSnitch.Domain.Models
 {
     public class Pm10Particle : IAirPollutionParticle
     {
-        public Pm10Particle(decimal value)
+        public Pm10Particle(double value)
         {
             Value = value;
         }
         public string ParticleName => "PM10";
-        public decimal Value { get; }
+        public double Value { get; }
     }
 }
